Move the StreamIndenter into pyomo.common.formatting
